// src/models/types/tv-db-response/tv-db-response-result.type.ts
export type TvDbResponseResult = {
  adult: boolean;
  backdrop_path: string;
  genre_ids: Array<number>;
  id: number;
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  first_air_date: string;
  name: string;
  video: false;
  vote_average: number;
  vote_count: number;
};